# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Appointment Scheduling API",
        "description": "This is a simple API for managing appointments.",
        "contact": {
            "email": "support@example.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/patients/{patientId}/appointments/upcomings": {
            "get": {
                "tags": [
                    "Patient Appointments"
                ],
                "summary": "Get patient's upcoming appointments",
                "description": "Fetch the upcoming appointments for a specific patient by patient ID. The patient's token must be valid.",
                "operationId": "718f65410611ee27774583bb737dff3c",
                "parameters": [
                    {
                        "name": "patientId",
                        "in": "path",
                        "description": "The ID of the patient",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "Authorization": {
                                        "description": "Bearer token for authentication",
                                        "type": "string",
                                        "example": "Bearer token"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "List of upcoming appointments for the patient",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "appointments": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "appointment_type": {
                                                        "type": "string",
                                                        "example": "follow_up"
                                                    },
                                                    "appointment_name": {
                                                        "type": "string",
                                                        "example": "Follow-up Consultation"
                                                    },
                                                    "start_time": {
                                                        "type": "string",
                                                        "format": "date-time",
                                                        "example": "2025-04-03T09:00:00"
                                                    },
                                                    "end_time": {
                                                        "type": "string",
                                                        "format": "date-time",
                                                        "example": "2025-04-03T09:30:00"
                                                    },
                                                    "doctor_id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "doctor_name": {
                                                        "type": "string",
                                                        "example": "Dr. Harry Potter"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Unauthorized, invalid or missing token",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Unauthorized"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/patients/{patientId}/appointments": {
            "post": {
                "tags": [
                    "Patient Appointments"
                ],
                "summary": "Create a new appointment for a patient",
                "description": "This endpoint allows the creation of a new appointment for the specified patient. The patient's token must be valid.",
                "operationId": "1e8ef864d52731bece0ebbd1923daaee",
                "parameters": [
                    {
                        "name": "patientId",
                        "in": "path",
                        "description": "The ID of the patient",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "patient_id": {
                                        "description": "The ID of the patient",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "doctor_id": {
                                        "description": "The ID of the doctor for the appointment",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "appointment_type": {
                                        "description": "The type of appointment",
                                        "type": "string",
                                        "example": "follow_up"
                                    },
                                    "start_time": {
                                        "description": "The start time of the appointment",
                                        "type": "string",
                                        "format": "date-time",
                                        "example": "2025-04-03T09:00:00"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Appointment created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Appointment created successfully"
                                        },
                                        "appointment": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "patient_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "doctor_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "appointment_type": {
                                                    "type": "string",
                                                    "example": "follow_up"
                                                },
                                                "start_time": {
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "example": "2025-04-03T09:00:00"
                                                },
                                                "end_time": {
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "example": "2025-04-03T09:30:00"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request due to invalid data",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Validation failed"
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Unauthorized, invalid or missing token",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Unauthorized"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/patients/{patientId}/appointments/{appointmentId}": {
            "delete": {
                "tags": [
                    "Patient Appointments"
                ],
                "summary": "Delete an appointment for a patient",
                "description": "Deletes an upcoming appointment for a patient if it exists and hasn’t already passed. Requires a valid Bearer token.",
                "operationId": "390d6f9bd28bd3ce3360690009379f3b",
                "parameters": [
                    {
                        "name": "patientId",
                        "in": "path",
                        "description": "The ID of the patient",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "appointmentId",
                        "in": "path",
                        "description": "The ID of the appointment to delete",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 10
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Appointment deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Appointment deleted successfully."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Cannot delete past appointments",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Cannot delete past appointments."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Unauthorized access",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Unauthorized"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Appointment not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Appointment not found."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}